+++ src/lightly_train/_commands/train_helpers.py
@@ -130,9 +130,10 @@ def get_dataloader(
     )
     if loader_args is not None:
         logger.debug(f"Using additional dataloader arguments {loader_args}.")
-        # Ignore batch_size from loader_args. It is already handled in
-        # get_global_batch_size.
+        # Ignore batch_size and num_workers from loader_args. They are already
+        # handled through dedicated function arguments.
         loader_args.pop("batch_size", None)
+        loader_args.pop("num_workers", None)
         dataloader_kwargs.update(**loader_args)
     return DataLoader(**dataloader_kwargs)
 
@@ -289,13 +290,16 @@ def get_optimizer_args(
     optim_type: OptimizerType | Literal["auto"],
     optim_args: dict[str, Any] | OptimizerArgs | None,
     method_cls: type[Method],
+    wrapped_model: ModelWrapper,
 ) -> OptimizerArgs:
     if isinstance(optim_args, OptimizerArgs):
         return optim_args
     optim_args = {} if optim_args is None else optim_args
     optim_args_cls = method_cls.optimizer_args_cls(optim_type=optim_type)
     logger.debug(f"Using optimizer '{optim_args_cls.type()}'.")
-    return validate.pydantic_model_validate(optim_args_cls, optim_args)
+    args = validate.pydantic_model_validate(optim_args_cls, optim_args)
+    args.resolve_auto(wrapped_model=wrapped_model)
+    return args
 
 
 def get_dataset_size(