+++ src/lightly_train/_models/torchvision/shufflenet.py
@@ -10,11 +10,16 @@ from torch import Tensor
 from torch.nn import Module
 from torchvision.models import ShuffleNetV2
 
-from lightly_train._models.model_wrapper import ForwardFeaturesOutput, ForwardPoolOutput
+from lightly_train._models.model_wrapper import (
+    ArchitectureInfo,
+    ArchitectureInfoGettable,
+    ForwardFeaturesOutput,
+    ForwardPoolOutput,
+)
 from lightly_train._models.torchvision.torchvision import TorchvisionModelWrapper
 
 
-class ShuffleNetV2ModelWrapper(TorchvisionModelWrapper):
+class ShuffleNetV2ModelWrapper(TorchvisionModelWrapper, ArchitectureInfoGettable):
     _torchvision_models = [ShuffleNetV2]
     _torchvision_model_name_pattern = r"shufflenet_v2.*"
 
@@ -40,3 +45,6 @@ class ShuffleNetV2ModelWrapper(TorchvisionModelWrapper):
     def feature_dim(self) -> int:
         feature_dim: int = self._model.fc.in_features  # type: ignore
         return feature_dim
+
+    def architecture_info(self) -> ArchitectureInfo:
+        return {"model_type": "convolutional", "norm_type": "batchnorm"}