+++ CHANGELOG.md
@@ -9,7 +9,16 @@ this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ### Added
 
+- Add `CopyBlend` augmentation for LTDETR object detection training.
+- Add logging of 1-based `epoch` to the console and the loggers.
+- Add `MixUp` augmentation for LTDETR object detection training.
+- Add support for COCO object detection dataset format.
+
 ### Changed
+
+- Make `ScaleJitter` in LTDETR step-aware. Now you can stop the augmentation by adding a
+  `step_stop` args like the following
+  `transform_args={"scale_jitter": {"step_stop": 10000}}`
 
 ### Deprecated